refactor: CAFETERIA 상수를 constants/cafeteria.ts로 분리
- utils/menu-policy.ts: 함수(isNextWeek, isNextWeekPublished, isAfterClose)만 유지
- constants/cafeteria.ts: CAFETERIA_OPEN_MIN, CAFETERIA_CLOSE_MIN, CAFETERIA_LABEL 이동
