+++ src/components/home/HeroStatus/hero-status.ts
@@ -3,7 +3,7 @@ import type {ElementType} from 'react';
 import {Clock, Moon, Sun, Utensils} from 'lucide-react';
 
 import dayjs from '@/lib/dayjs';
-import {CAFETERIA_CLOSE_MIN, CAFETERIA_LABEL, CAFETERIA_OPEN_MIN} from '@/utils/menu-policy';
+import {CAFETERIA_CLOSE_MIN, CAFETERIA_LABEL, CAFETERIA_OPEN_MIN} from '@/constants/cafeteria';
 import {formatYYYYMMDD} from '@/utils/utils';
 import {MenuType} from '@/types/menu';
